Added RoomsController
- Added IRoomsService interface and RoomsService class with implemented CRUD operations
- Added ErrorHandlerMiddleware
- Improved UserService's UpdateAsync and DeleteAsync methods

+++ src/Business/Services/UsersService.cs
@@ -1,13 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Castle.Core.Internal;
-using Castle.DynamicProxy.Generators.Emitters.SimpleAST;
+using Castle.DynamicProxy.Contributors;
 using HotelReservation.Business.Interfaces;
 using HotelReservation.Business.Models.UserModels;
 using HotelReservation.Data.Entities;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 
 namespace HotelReservation.Business.Services
@@ -16,17 +18,14 @@ namespace HotelReservation.Business.Services
     {
         private readonly UserManager<UserEntity> _userManager;
         private readonly IMapper _mapper;
-        private readonly RoleManager<IdentityRole> _roleManager;
         private readonly IPasswordHasher<UserEntity> _passwordHasher;
 
         public UsersService(
             UserManager<UserEntity> userManager,
-            RoleManager<IdentityRole> roleManger,
             IPasswordHasher<UserEntity> passwordHasher,
             IMapper mapper)
         {
             _userManager = userManager;
-            _roleManager = roleManger;
             _passwordHasher = passwordHasher;
             _mapper = mapper;
         }
@@ -45,12 +44,12 @@ namespace HotelReservation.Business.Services
         public async Task<UserModel> CreateAsync(UserRegistrationModel userRegistration)
         {
             if (userRegistration == null)
-                throw new DataException("User cannot be empty", ErrorStatus.EmptyInput);
+                throw new BusinessException("User cannot be empty", ErrorStatus.EmptyInput);
 
             var existingUserEntity = await _userManager.FindByEmailAsync(userRegistration.Email);
 
             if (existingUserEntity != null)
-                throw new DataException("User with such email already exists", ErrorStatus.AlreadyExist);
+                throw new BusinessException("User with such email already exists", ErrorStatus.AlreadyExist);
 
             var userEntity = _mapper.Map<UserEntity>(userRegistration);
 
@@ -85,26 +84,34 @@ namespace HotelReservation.Business.Services
             return userModel;
         }
 
-        public async Task<UserModel> DeleteAsync(string id)
+        public async Task<UserModel> DeleteAsync(string id, IEnumerable<Claim> currentUserClaims)
         {
+            // could be removed
+            if (currentUserClaims.FirstOrDefault(cl => cl.Type == "id").Value == id)
+                throw new BusinessException("You cannot delete yourself", ErrorStatus.IncorrectInput);
+
             var deletedUserEntity = await _userManager.FindByIdAsync(id);
+
             var deletedUserModel = _mapper.Map<UserModel>(deletedUserEntity);
             await GetRolesForUserModelAsync(deletedUserModel);
 
             var result = await _userManager.DeleteAsync(deletedUserEntity);
 
+            if (result != IdentityResult.Success)
+                throw new BusinessException("No user with such id", ErrorStatus.NotFound);
+
             return deletedUserModel;
         }
 
-        public async Task<UserModel> UpdateAsync(string id, UserUpdateModel userUpdateModel)
+        public async Task<UserModel> UpdateAsync(string id, UserUpdateModel userUpdateModel, IEnumerable<Claim> currentUserClaims)
         {
             if (userUpdateModel == null)
-                throw new DataException("User cannot be empty", ErrorStatus.EmptyInput);
+                throw new BusinessException("User cannot be empty", ErrorStatus.EmptyInput);
 
             var userEntity = await _userManager.FindByIdAsync(id);
 
             if (userEntity == null)
-                throw new DataException("User with such id does not exist", ErrorStatus.NotFound);
+                throw new BusinessException("User with such id does not exist", ErrorStatus.NotFound);
 
             if (userUpdateModel.Email != null)
                 userEntity.Email = userUpdateModel.Email;
@@ -134,35 +141,48 @@ namespace HotelReservation.Business.Services
                     userUpdateModel.NewPassword);
 
                 if (changePasswordResult != IdentityResult.Success)
-                    throw new DataException("Incorrect previous password", ErrorStatus.IncorrectInput);
+                    throw new BusinessException("Incorrect previous password", ErrorStatus.IncorrectInput);
             }
 
             var result = await _userManager.UpdateAsync(userEntity);
 
             var updatedUserEntity = await _userManager.FindByEmailAsync(userEntity.Email);
 
-            var addedUserRoles = await _userManager.GetRolesAsync(updatedUserEntity);
-
             if (result == IdentityResult.Success)
             {
-                if (!addedUserRoles.Contains("User"))
-                    await _userManager.AddToRoleAsync(updatedUserEntity, "User");
-
-                for (int i = 0; i < userUpdateModel.Roles.Count(); ++i)
+                if (userUpdateModel.Roles != null)
                 {
-                    userUpdateModel.Roles[i] = userUpdateModel.Roles[i].ToUpper();
-                }
+                    var addedUserRoles = await _userManager.GetRolesAsync(updatedUserEntity);
 
-                foreach (var role in addedUserRoles)
-                {
-                    if (!userUpdateModel.Roles.Contains(role.ToUpper()))
-                        await _userManager.RemoveFromRoleAsync(updatedUserEntity, role.ToUpper());
-                }
+                    if (!addedUserRoles.Contains("User"))
+                        await _userManager.AddToRoleAsync(updatedUserEntity, "User");
 
-                foreach (var role in userUpdateModel.Roles)
-                {
-                    if (role.ToUpper() == "ADMIN" || role.ToUpper() == "MANAGER")
-                        await _userManager.AddToRoleAsync(updatedUserEntity, role);
+                    for (int i = 0; i < userUpdateModel.Roles.Count(); ++i)
+                    {
+                        userUpdateModel.Roles[i] = userUpdateModel.Roles[i].ToUpper();
+                    }
+
+                    foreach (var role in addedUserRoles)
+                    {
+                        if (!userUpdateModel.Roles.Contains(role.ToUpper()))
+                        {
+                            if (role.ToUpper() == "ADMIN" &&
+                                updatedUserEntity.Id == currentUserClaims.FirstOrDefault(cl => cl.Type == "id").Value)
+                            {
+                                throw new BusinessException(
+                                    "You cannot change your own admin role",
+                                    ErrorStatus.IncorrectInput);
+                            }
+
+                            await _userManager.RemoveFromRoleAsync(updatedUserEntity, role.ToUpper());
+                        }
+                    }
+
+                    foreach (var role in userUpdateModel.Roles)
+                    {
+                        // if (role.ToUpper() == "ADMIN" || role.ToUpper() == "MANAGER")
+                            await _userManager.AddToRoleAsync(updatedUserEntity, role);
+                    }
                 }
             }
